+++ src/components/atoms/TextInput/TextInput.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import { TextInput } from './TextInput';
 
 const meta: Meta<typeof TextInput> = {
-  title: 'Atoms/TextInput',
+  title: 'Atoms/Inputs/TextInput',
   component: TextInput,
 };