+++ src/components/index.ts
@@ -2,3 +2,4 @@ export * from './atoms';
 export * from './molecules';
 export * from './auth';
 export * from './form';
+export * from './organisms';