+++ src/locales/cs/common.ts
@@ -1,6 +1,7 @@
 export const common = {
   forward: 'Dopředu',
   back: 'Zpět',
+  close: 'Zavřít',
   'languages.english': 'Angličtina',
   'languages.czech': 'Čeština',
   'theme.system': 'Systém',