+++ src/utils/hooks/index.ts
@@ -1,5 +1,4 @@
 export * from './useAuth';
 export * from './useRequest';
 export * from './storage';
-export * from './useCommonTranslation';
 export * from './useFormController';